+++ market-data/src/main/java/com/multiassetoms/marketdata/model/MarketPriceNotFoundException.java
@@ -1,0 +1,8 @@
+package com.multiassetoms.marketdata.model;
+
+public class MarketPriceNotFoundException extends RuntimeException {
+
+    public MarketPriceNotFoundException(String message) {
+        super(message);
+    }
+}